- Moved the one-to-one creation form and view from events/ to people/
