+++ events/forms.py
@@ -32,13 +32,3 @@ class WorkshopForm(forms.Form):
             "startTime",
             "duration",
         )
-
-
-class CreateMeetingForm(forms.Form):
-    name = forms.CharField(widget=forms.Textarea)
-    start_time = forms.DateTimeField()
-    location = forms.CharField(widget=forms.Textarea)
-    duration = forms.IntegerField()
-
-    # Change this to be all mentors with a relationship to the mentee.
-    mentor = forms.ModelChoiceField(queryset=User.objects.all())
